Implement a UDP topic injector
This is a feature useful for Hardware-In-The-Loop (HITL) testing. It
allows a remote computer to send messages to the board and they will be
published on the message bus as if they were generated locally. It can
be used for testing the software without having the full boat, for
continuous integration or to prototype algorithms on a PC before moving
to a "real" implementation in C.